requirements.txt
discord.py==2.3.2
crewai
pyyaml
python-dotenv
requests
